+++ frontend/src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Navigation from './components/Navigation';
+import AddTrip from './components/AddTrip';
 
 function App() {
   return (
@@ -12,6 +13,7 @@ function App() {
         </Route>
         <Route path="/add">
           <p>Add trip page</p>
+          <AddTrip />
         </Route>
         <Route path="/login">
           <p>Login page</p>
